Fixed bug on si7020 C driver, now Temperature and humidity are correctly reads

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/network/Protocol.java
@@ -57,7 +57,7 @@ public abstract class Protocol
     		if(bestPath >= 0){
     			System.out.print("send unicast to ");
     			System.out.println(bestPath);
-    			Radio.send((short)bestPath, p.toByteArray());     
+    			Radio.send(bestPath, p.toByteArray());     
     		}else{
     			System.out.println("send broadcast");
     			Radio.broadcast(p.toByteArray());
